# documentation/openapi.yml
openapi : 3.0.3
info :
    description : This is the specification for the Stadtgemüse Service.
    version : 2.0.0
    title : Stadtgemüse Service
    contact : { }
servers :
    -   url : 'https://veg-service.herokuapp.com/v2/'
        description : Production v2
...
paths :
...
    '/purchase' :
        get :
            security :
                -   bearerAuth : [ ]
            tags :
                - Purchase
            summary : Retrieve all previous purchases
            operationId : purchaseGet
            responses :
                '200' :
                    description : Successful call, returns the list of all previous purchases
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/purchase-list-response'
                default :
                    description : Information about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
        post :
            security :
                -   bearerAuth : [ ]
            tags :
                - Purchase
            summary : Purchases items for a user. Reduces the amount of these items in stock and the users balance
            operationId : purchaseFromStock
            requestBody :
                description : All items that the user wants to purchase
                required : true
                content :
                    application/json :
                        schema :
                            $ref : '#/components/schemas/purchase-request'
            responses :
                '200' :
                    description : Successful call, returns the price of what was purchase and the new balance
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/purchase-response'
                default :
                    description : Details about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
    '/purchase/{purchase-id}' :
        get :
            security :
                -   bearerAuth : [ ]
            tags :
                - Purchase
            summary : Retrieve one purchase by Id
            operationId : purchaseGetById
            parameters :
                -   $ref : '#/components/parameters/purchase-id'
            responses :
                '200' :
                    description : Successful call, returns a whole purchase
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/purchase-history-item'
                default :
                    description : Information about the error that occured
                    content :
                        application/json :
                            schema :
                                $ref : '#/components/schemas/error-response'
...
components :
    securitySchemes :
        bearerAuth : # arbitrary name for the security scheme
            type : http
            scheme : bearer
            bearerFormat : JWT
    schemas :
...
        error-response :
            description : "Contains details about the occurred error"
            type : object
            required :
                - errorCode
                - status
                - errorMessage
            properties :
                errorCode :
                    description : "Number, that can be referenced to a specific error"
                    example : 400001
                    type : integer
                status :
                    description : "HTTP Status Code returned"
                    example : 400
                    type : integer
                message :
                    description : "A detailed error message"
                    example : "Validation failed"
                    type : string
                details :
                    type : array
                    items :
                        $ref : '#/components/schemas/error-detail'
        error-detail :
            description : A more precise explaination of an error
            type : object
            properties :
                message :
                    description : A message relating to the error detail
                    type : string
                    example : "Invalid request parameter: balance"
                detail :
                    description : A detailed message of the error
                    type : string
                    example : "The value null does not match the Specification"

        #
        # Request Schemas
        #
...
        purchase-request :
            description : All items that the user wants to purchase
            type : object
            required :
                - items
            properties :
                items :
                    type : array
                    minItems : 1
                    items :
                        $ref : '#/components/schemas/cart-item'
...
        purchase-response :
            description : Information about the users balance and the purchased cart
            type : object
            properties :
                name :
                    $ref : '#/components/schemas/username'
                id :
                    $ref : '#/components/schemas/uuid'
                balance :
                    description : The balance of the user after the purchase
                    type : number
                    format : float
                    example : 140.8
                price :
                    description : How much the cart cost
                    type : number
                    format : float
                    example : 17.43
...
        purchase-list-response :
            description : List of all previous purchases
            type : object
            properties :
                purchases :
                    type : array
                    items :
                        $ref : '#/components/schemas/purchase-history-item'
        purchase-history-item :
            description : Contains information about a whole purchase
            type : object
            properties :
                id :
                    $ref : '#/components/schemas/uuid'
                createdOn :
                    description : Date and time the purchase was created at
                    example : '2017-07-21T17:32:28Z'
                    type : string
                    format : date-time
                totalPrice :
                    description : |-
                        Total price of the purchase.
                        Changes to an items price after the purchase was made will not be reflected in this value.
                    example : 69.0
                    type : number
                    format : float
                items :
                    type : array
                    items :
                        $ref : '#/components/schemas/purchase-item'
        purchase-item :
            description : Contains information about a single item that was purchased
            type : object
            properties :
                id :
                    $ref : '#/components/schemas/uuid'
                name :
                    $ref : '#/components/schemas/item-name'
                amount :
                    description : Amount of items purchased
                    example : 4.2
                    type : number
                    format : float
                pricePerUnit :
                    description : |-
                        Price per unit of the item.
                        Changes to an items price after the purchase was made will not be reflected in this value.
                    example : 12.0
                    type : number
                    format : float
                unitType :
                    $ref : '#/components/schemas/unitType'
...
        uuid :
            description : Unique identifier
            example : "0794e2e2-f5b2-42fb-a2b3-e12c75897f85"
            type : string
        item-name :
            description : Name of the stock item
            example : Kartoffeln
            type : string
            minLength : 1
            maxLength : 250
        unitType :
            description : |-
                By which measurement the item is sold
                  * WEIGHT: Kilogram, or fractions of a kilogram
                  * PIECE: Positive amount of whole pieces
            example : PIECE
            type : string
            enum :
                - PIECE
                - WEIGHT
...
        cart-item :
            description : An specific item including its quantity the user wants to purchase
            type : object
            required :
                - id
                - amount
            properties :
                id :
                    $ref : '#/components/schemas/uuid'
                amount :
                    type : number
                    description : |-
                        How many actual items the user wants to buy.
                        Must be an integer if the unitType of the referenced item is PIECE
                    format : float
                    example : 7
        username :
            description : The unique name of the user
            type : string
            example : Robby
            maxLength : 255
...
    parameters :
...
        purchase-id :
            in : path
            name : purchase-id
            schema :
                type : string
                maxLength : 36
            required : true
            description : The unique id of a purchase